+++ notes/src/Components/Note/Note.css
