+++ notes/src/Components/NoteCard/NoteCard.css
