+++ notes/src/Components/NoteCard/NoteCard.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Menu from './Menu';
+import Menu from '../Menu/Menu';
 import { Link } from 'react-router-dom';
 import './NoteCard.css';
